refactor:불변하는 변수 키워드 const로 변경
